Synchronize with master branch of private code
[issue#] N/A
[cause] N/A
[solution] Synchronize with master branch of private code
Changes)
- Apply latest UI and GUI guide
- Update po files
- Resolve a bug for Oragne TC
- Remove unnecessary dlog messages
- Default view modie is changed to hidden view

	modified:   CMakeLists.txt
	modified:   oma-parser/oma-parser-common.h
	modified:   oma-parser/oma-parser-dd1.c
	modified:   oma-parser/oma-parser-dd1.h
	modified:   po/ar.po
	modified:   po/az.po
	modified:   po/bg.po
	modified:   po/ca.po
	modified:   po/cs.po
	modified:   po/da.po
	modified:   po/de_DE.po
	modified:   po/el_GR.po
	modified:   po/es_ES.po
	modified:   po/es_MX.po
	modified:   po/et.po
	modified:   po/eu.po
	modified:   po/fi.po
	modified:   po/fr_CA.po
	modified:   po/fr_FR.po
	modified:   po/ga.po
	modified:   po/gl.po
	modified:   po/hi.po
	modified:   po/hr.po
	modified:   po/hu.po
	modified:   po/hy.po
	modified:   po/is.po
	modified:   po/it_IT.po
	modified:   po/ja_JP.po
	modified:   po/ka.po
	modified:   po/kk.po
	modified:   po/ko_KR.po
	modified:   po/lt.po
	modified:   po/lv.po
	modified:   po/mk.po
	modified:   po/nb.po
	modified:   po/nl_NL.po
	modified:   po/pl.po
	modified:   po/pt_BR.po
	modified:   po/pt_PT.po
	modified:   po/ro.po
	modified:   po/ru_RU.po
	modified:   po/sk.po
	modified:   po/sl.po
	modified:   po/sr.po
	modified:   po/sv.po
	modified:   po/tr_TR.po
	modified:   po/uk.po
	modified:   po/uz.po
	modified:   po/zh_CN.po
	modified:   po/zh_HK.po
	modified:   po/zh_SG.po
	modified:   po/zh_TW.po
	renamed:    res/images/Q02_Notification_Download_failed.png -> res/images/00_icon_delete.png
	new file:   res/images/Notification_download_animation.gif
	modified:   res/images/Q02_Notification_Download_failed.png
	modified:   res/images/Q02_Notification_download_complete.png
	renamed:    res/images/Q02_Notification_download_complete.png -> res/images/U01_icon_broken.png
	renamed:    res/images/file_lock_03.png -> res/images/U01_icon_drm.png
	new file:   res/images/U01_icon_excel.png
	new file:   res/images/U01_icon_html.png
	new file:   res/images/U01_icon_pdf.png
	new file:   res/images/U01_icon_ppt.png
	new file:   res/images/U01_icon_swf.png
	new file:   res/images/U01_icon_text.png
	new file:   res/images/U01_icon_tpk.png
	new file:   res/images/U01_icon_unkown.png
	new file:   res/images/U01_icon_vcs.png
	new file:   res/images/U01_icon_word.png
	deleted:    res/images/U01_list_icon_Broken.png
	deleted:    res/images/U01_list_icon_ETC.png
	deleted:    res/images/U01_list_icon_PDF.png
	deleted:    res/images/U01_list_icon_Vcalendar.png
	deleted:    res/images/U01_list_icon_excel.png
	deleted:    res/images/U01_list_icon_flash.png
	deleted:    res/images/U01_list_icon_html.png
	modified:   res/images/U01_list_icon_image.png
	modified:   res/images/U01_list_icon_mp3.png
	new file:   res/images/U01_list_icon_mp4.png
	deleted:    res/images/U01_list_icon_powerpoint.png
	deleted:    res/images/U01_list_icon_text.png
	deleted:    res/images/U01_list_icon_tpk_file.png
	deleted:    res/images/U01_list_icon_video.png
	deleted:    res/images/U01_list_icon_word-file.png
	modified:   src/download-manager-dateTime.cpp
	modified:   src/download-manager-downloadItem.cpp
	modified:   src/download-manager-event.cpp
	modified:   src/download-manager-history-db.cpp
	modified:   src/download-manager-item.cpp
	modified:   src/download-manager-network.cpp
	modified:   src/download-manager-notification.cpp
	modified:   src/download-manager-util.cpp
	modified:   src/download-manager-view.cpp
	modified:   src/download-manager-viewItem.cpp
	modified:   src/include/download-manager-common.h
	modified:   src/include/download-manager-debug.h
	modified:   src/include/download-manager-downloadItem.h
	modified:   src/include/download-manager-item.h
	modified:   src/include/download-manager-items.h
	modified:   src/include/download-manager-notification.h
	modified:   src/include/download-manager-util.h
	modified:   src/include/download-manager-view.h
	modified:   src/include/download-manager-viewItem.h
	modified:   src/main.cpp

+++ oma-parser/oma-parser-dd1.c
@@ -377,11 +377,11 @@ void op_expat_character_dd1(void *userData, const XML_Char *s, int len)
 
 			op_com_util_remove_blankspace(ch_str, len);
 			if (op_com_utils_is_number((const char*)ch_str)) {
-				dd_info->size = atoi(ch_str);
+				dd_info->size = atoll(ch_str);
 				if (dd_info->size <= 0)
 					OP_LOGE("size is invalid");
 
-				OP_SLOG("dd_info->size:[%d]", dd_info->size);
+				OP_SLOG("dd_info->size:[%llu]", dd_info->size);
 			} else {
 				OP_LOGE("size is invalid");
 			}
